refactored repository package
